Move service context code in a separate file
